Create process_client_logs.sh
The script as of now can detect all files in an unzipped log directory.
It then adds all of the logs accross all of the log files into one `result.txt`.
We now have to figure out how to refine the regex to be specific to what we want,
and not literally everything.