Create `create_country_dist.sh`
We use our most succinct REGEX to grab the IP from each
`failed_login_data.txt`. That is sorted and held in an intermediate
`ip.text` that is deleted at the end. That `ip.text` is joined
with `etc/country_IP_map.txt`. We use an awk statement to swap
the IP and country from that join so that we can sort by country
to get unique countries. Then we get the unique count of coutnries and
wrap it all up as desired.